fix(synthetix): extracted getting synthetix's holder into a helper and fixed GQL errors
